Remove io_client directory
