+++ adapter/filesystem/filesystem_mdm.h
@@ -15,7 +15,7 @@
 
 #include <cstdio>
 #include <unordered_map>
-#include "io_client/filesystem/filesystem_io_client.h"
+#include "filesystem_io_client.h"
 #include "filesystem.h"
 #include "thread_pool.h"
