+++ adapter/mpiio/mpiio_api.cc
@@ -15,7 +15,7 @@ bool mpiio_intercepted = true;
 #include <hermes.h>
 #include <bucket.h>
 
-#include "io_client/mpiio/mpiio_api.h"
+#include "mpiio_api.h"
 #include "mpiio_fs_api.h"
 
 #include "hermes_shm/util/singleton.h"
