+++ adapter/mpiio/mpiio_io_client.cc
