+++ adapter/mpiio/mpiio_io_client.h
@@ -10,52 +10,53 @@
  * have access to the file, you may request a copy from help@hdfgroup.org.   *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-#ifndef HERMES_ADAPTER_STDIO_STDIO_IO_CLIENT_H_
-#define HERMES_ADAPTER_STDIO_STDIO_IO_CLIENT_H_
+#ifndef HERMES_ADAPTER_MPIIO_MPIIO_IO_CLIENT_H_
+#define HERMES_ADAPTER_MPIIO_MPIIO_IO_CLIENT_H_
 
 #include <memory>
 
-#include "io_client/filesystem/filesystem_io_client.h"
-#include "stdio_api.h"
+#include "adapter/filesystem/filesystem_io_client.h"
+#include "mpiio_api.h"
 
 using hermes::adapter::fs::AdapterStat;
 using hermes::adapter::fs::FsIoOptions;
 using hermes::adapter::fs::IoStatus;
-using hermes::adapter::fs::StdioApi;
+using hermes::adapter::fs::MpiioApi;
 
 namespace hermes::adapter::fs {
 
-/** State for the STDIO I/O trait */
-struct StdioIoClientHeader : public TraitHeader {
-  explicit StdioIoClientHeader(const std::string &trait_uuid,
+/** State for the MPI I/O trait */
+struct MpiioIoClientHeader : public TraitHeader {
+  explicit MpiioIoClientHeader(const std::string &trait_uuid,
                                const std::string &trait_name)
-      : TraitHeader(trait_uuid, trait_name, HERMES_TRAIT_FLUSH) {}
+      : TraitHeader(trait_uuid, trait_name,
+                    HERMES_TRAIT_FLUSH) {}
 };
 
 /** A class to represent STDIO IO file system */
-class StdioIoClient : public hermes::adapter::fs::FilesystemIoClient {
- public:
-  HERMES_TRAIT_H(StdioIoClient, "stdio_io_client")
+class MpiioIoClient : public hermes::adapter::fs::FilesystemIoClient {
+public:
+  HERMES_TRAIT_H(MpiioIoClient, "mpiio_io_client")
 
  private:
-  HERMES_STDIO_API_T real_api; /**< pointer to real APIs */
+  HERMES_MPIIO_API_T real_api; /**< pointer to real APIs */
 
  public:
   /** Default constructor */
-  StdioIoClient() {
-    real_api = HERMES_STDIO_API;
-    CreateHeader<StdioIoClientHeader>(trait_name_, trait_name_);
+  MpiioIoClient() {
+    real_api = HERMES_MPIIO_API;
+    CreateHeader<MpiioIoClientHeader>("mpiio_io_client_", trait_name_);
   }
 
   /** Trait deserialization constructor */
-  explicit StdioIoClient(hshm::charbuf &params) {
+  explicit MpiioIoClient(hshm::charbuf &params) {
     (void) params;
-    real_api = HERMES_STDIO_API;
-    CreateHeader<StdioIoClientHeader>(trait_name_, trait_name_);
+    real_api = HERMES_MPIIO_API;
+    CreateHeader<MpiioIoClientHeader>("mpiio_io_client_", trait_name_);
   }
 
   /** Virtual destructor */
-  virtual ~StdioIoClient() = default;
+  virtual ~MpiioIoClient() = default;
 
  public:
   /** Allocate an fd for the file f */
@@ -95,6 +96,11 @@ class StdioIoClient : public hermes::adapter::fs::FilesystemIoClient {
   /** Get initial statistics from the backend */
   size_t GetSize(const hipc::charbuf &bkt_name) override;
 
+  /** Initialize I/O context using count + datatype */
+  static size_t IoSizeFromCount(int count,
+                                MPI_Datatype datatype,
+                                FsIoOptions &opts);
+
   /** Write blob to backend */
   void WriteBlob(const std::string &bkt_name,
                  const Blob &full_blob,
@@ -106,13 +112,16 @@ class StdioIoClient : public hermes::adapter::fs::FilesystemIoClient {
                 Blob &full_blob,
                 const FsIoOptions &opts,
                 IoStatus &status) override;
+
+  /** Update the I/O status after a ReadBlob or WriteBlob */
+  void UpdateIoStatus(size_t count, IoStatus &status);
 };
 
 }  // namespace hermes::adapter::fs
 
 /** Simplify access to the stateless StdioIoClient Singleton */
-#define HERMES_STDIO_IO_CLIENT \
-  hshm::EasySingleton<hermes::adapter::fs::StdioIoClient>::GetInstance()
-#define HERMES_STDIO_IO_CLIENT_T hermes::adapter::fs::StdioIoClient*
+#define HERMES_MPIIO_IO_CLIENT \
+  hshm::EasySingleton<hermes::adapter::fs::MpiioIoClient>::GetInstance()
+#define HERMES_MPIIO_IO_CLIENT_T hermes::adapter::fs::MpiioIoClient*
 
-#endif  // HERMES_ADAPTER_STDIO_STDIO_IO_CLIENT_H_
+#endif  // HERMES_ADAPTER_MPIIO_MPIIO_IO_CLIENT_H_
