+++ adapter/posix/posix_api_singleton.cc
