+++ adapter/posix/posix_io_client.cc
