+++ adapter/stdio/CMakeLists.txt
@@ -6,6 +6,19 @@ include_directories(
         ${HERMES_IO_CLIENT_DIR}
         ${CMAKE_CURRENT_SOURCE_DIR})
 
+# Creates the STDIO I/O singleton
+add_library(hermes_stdio_io_real stdio_api_singleton.cc)
+target_link_libraries(hermes_stdio_io_real stdc++fs dl)
+
+# Creates the STDIO I/O client
+add_library(hermes_stdio_io_client stdio_io_client.cc)
+target_compile_options(hermes_stdio_io_client PUBLIC -fPIC)
+add_dependencies(hermes_stdio_io_client
+        hermes hermes_stdio_io_real)
+target_link_libraries(hermes_stdio_io_client
+        hermes hermes_stdio_io_real
+        MPI::MPI_CXX stdc++fs dl)
+
 # Create the STDIO interceptor
 set(INTERCEPTOR_DEPS
         hermes
@@ -22,6 +35,8 @@ target_compile_options(hermes_stdio PUBLIC -fPIC)
 #-----------------------------------------------------------------------------
 install(
         TARGETS
+        hermes_stdio_io_real
+        hermes_stdio_io_client
         hermes_stdio
         EXPORT
         ${HERMES_EXPORTED_TARGETS}
@@ -29,9 +44,29 @@ install(
         ARCHIVE DESTINATION ${HERMES_INSTALL_LIB_DIR}
         RUNTIME DESTINATION ${HERMES_INSTALL_BIN_DIR}
 )
+
+#-----------------------------------------------------------------------------
+# Export all exported targets to the build tree for use by parent project
+#-----------------------------------------------------------------------------
+set(HERMES_EXPORTED_LIBS
+        hermes_stdio_io_real
+        hermes_stdio_io_client
+        hermes_stdio
+        ${HERMES_EXPORTED_LIBS})
+if(NOT HERMES_EXTERNALLY_CONFIGURED)
+    EXPORT (
+            TARGETS
+            ${HERMES_EXPORTED_LIBS}
+            FILE
+            ${HERMES_EXPORTED_TARGETS}.cmake
+    )
+endif()
+
 #-----------------------------------------------------------------------------
 # Add Target(s) to Coverage
 #-----------------------------------------------------------------------------
 if(HERMES_ENABLE_COVERAGE)
+    set_coverage_flags(hermes_stdio_io_real)
+    set_coverage_flags(hermes_stdio_io_client)
     set_coverage_flags(hermes_stdio)
 endif()
