+++ adapter/stdio/stdio_api_singleton.cc
