+++ adapter/stdio/stdio_io_client.cc
