+++ adapter/test/posix/posix_adapter_mpi_test.cpp
@@ -22,7 +22,7 @@
 #include "adapter_test_utils.h"
 
 #if HERMES_INTERCEPT == 1
-#include "io_client/posix/posix_api.h"
+#include "adapter/posix/posix_api.h"
 #include "adapter/posix/posix_fs_api.h"
 #endif
