+++ adapter/test/stdio/stdio_adapter_mpi_test.cpp
@@ -19,7 +19,7 @@
 #include <filesystem>
 #include <iostream>
 #if HERMES_INTERCEPT == 1
-#include "io_client/stdio/stdio_api.h"
+#include "adapter/stdio/stdio_api.h"
 #include "adapter/stdio/stdio_fs_api.h"
 #endif
