--- io_client/posix/CMakeLists.txt
@@ -1,49 +1,0 @@
-# Creates the POSIX I/O singleton
-add_library(hermes_posix_io_real posix_api_singleton.cc)
-target_link_libraries(hermes_posix_io_real
-        stdc++fs dl)
-# Creates the POSIX I/O client and singleton
-add_library(hermes_posix_io_client
-        posix_io_client.cc)
-target_compile_options(hermes_posix_io_client PUBLIC -fPIC)
-add_dependencies(hermes_posix_io_client
-        hermes hermes_posix_io_real)
-target_link_libraries(hermes_posix_io_client
-        hermes hermes_posix_io_real
-        MPI::MPI_CXX stdc++fs dl)
-
-#-----------------------------------------------------------------------------
-# Add Target(s) to CMake Install
-#-----------------------------------------------------------------------------
-install(
-        TARGETS
-        hermes_posix_io_real
-        hermes_posix_io_client
-        EXPORT
-        ${HERMES_EXPORTED_TARGETS}
-        LIBRARY DESTINATION ${HERMES_INSTALL_LIB_DIR}
-        ARCHIVE DESTINATION ${HERMES_INSTALL_LIB_DIR}
-        RUNTIME DESTINATION ${HERMES_INSTALL_BIN_DIR}
-)
-#-----------------------------------------------------------------------------
-# Export all exported targets to the build tree for use by parent project
-#-----------------------------------------------------------------------------
-set(HERMES_EXPORTED_LIBS
-        hermes_posix_io_real
-        hermes_posix_io_client
-        ${HERMES_EXPORTED_LIBS})
-if(NOT HERMES_EXTERNALLY_CONFIGURED)
-    EXPORT (
-            TARGETS
-            ${HERMES_EXPORTED_LIBS}
-            FILE
-            ${HERMES_EXPORTED_TARGETS}.cmake
-    )
-endif()
-#-----------------------------------------------------------------------------
-# Add Target(s) to Coverage
-#-----------------------------------------------------------------------------
-if(HERMES_ENABLE_COVERAGE)
-    set_coverage_flags(hermes_posix_io_real)
-    set_coverage_flags(hermes_posix_io_client)
-endif()
