+++ src/borg_io_clients/borg_posix_client.h
@@ -14,7 +14,7 @@
 #define HERMES_SRC_BORG_IO_CLIENTS_POSIX_H
 
 #include "borg_io_client.h"
-#include "io_client/posix/posix_api.h"
+#include "adapter/posix/posix_api.h"
 
 #include <filesystem>
